tests/wpt/web-platform-tests/css/css-images/parsing/image-orientation-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Images: getComputedStyle().imageOrientation</title>
<link rel="help" href="https://drafts.csswg.org/css-images-3/#propdef-image-orientation">
<meta name="assert" content="image-orientation computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("image-orientation", "from-image");
test_computed_value("image-orientation", "none");
</script>
</body>
</html>
