src/third_party/blink/web_tests/external/wpt/css/css-images/parsing/image-orientation-computed-expected.txt

This is a testharness.js-based test.
PASS Property image-orientation value 'from-image' computes to 'from-image'
FAIL Property image-orientation value 'none' computes to 'none' assert_equals: expected "none" but got "0deg"
Harness: the test ran to completion.
